+++ server/party_handler_test.go
@@ -15,6 +15,7 @@
 package server
 
 import (
+	"context"
 	"testing"
 
 	"github.com/gofrs/uuid/v5"
@@ -64,7 +65,8 @@ func createTestPartyHandler(t *testing.T, logger *zap.Logger) (*PartyHandler, fu
 
 	dmr := DummyMessageRouter{}
 
-	pr := NewLocalPartyRegistry(logger, cfg, mm, &tt, &tsm, &dmr, node)
+	pr := NewLocalPartyRegistry(context.Background(), logger, logger, cfg, node)
+	pr.Init(mm, &tt, &tsm, &dmr)
 	ph := NewPartyHandler(logger, pr, mm, &tt, &tsm, &dmr, uuid.UUID{}, node, true, 10, nil)
 	return ph, cleanup
 }