+++ main.go
@@ -40,6 +40,7 @@ import (
 	"go.uber.org/zap/zapcore"
 	"google.golang.org/protobuf/encoding/protojson"
 
+	_ "github.com/felixge/httpsnoop"
 	_ "github.com/klauspost/compress"
 	_ "github.com/prometheus/client_golang/prometheus"
 	_ "github.com/prometheus/common/model"